+++ packages/web/src/javascripts/NativeMobileWeb/DownloadSelectedNotesOnAndroid.tsx
@@ -1,21 +1,20 @@
 import { WebApplication } from '@/Application/Application'
-import { getBase64FromBlob } from '@/Utils'
 import { getNoteBlob, getNoteFileName } from '@/Utils/NoteExportUtils'
 import { parseFileName } from '@standardnotes/filepicker'
-import { SNNote } from '@standardnotes/snjs'
+import { Platform, SNNote } from '@standardnotes/snjs'
 import { sanitizeFileName } from '@standardnotes/ui-services'
+import { downloadBlobOnAndroid } from './DownloadBlobOnAndroid'
 
-export const shareSelectedItems = async (application: WebApplication, notes: SNNote[]) => {
-  if (!application.isNativeMobileWeb()) {
-    throw new Error('Share function being used outside mobile webview')
+export const downloadSelectedNotesOnAndroid = async (application: WebApplication, notes: SNNote[]) => {
+  if (!application.isNativeMobileWeb() || application.platform !== Platform.Android) {
+    throw new Error('Function being used on non-android platform')
   }
   if (notes.length === 1) {
     const note = notes[0]
     const blob = getNoteBlob(application, note)
-    const base64 = await getBase64FromBlob(blob)
     const { name, ext } = parseFileName(getNoteFileName(application, note))
     const filename = `${sanitizeFileName(name)}.${ext}`
-    application.mobileDevice.shareBase64AsFile(base64, filename)
+    await downloadBlobOnAndroid(application, blob, filename)
     return
   }
   if (notes.length > 1) {
@@ -27,10 +26,7 @@ export const shareSelectedItems = async (application: WebApplication, notes: SNN
         }
       }),
     )
-    const zippedDataAsBase64 = await getBase64FromBlob(zippedDataBlob)
-    application.mobileDevice.shareBase64AsFile(
-      zippedDataAsBase64,
-      `Standard Notes Export - ${application.getArchiveService().formattedDateForExports()}.zip`,
-    )
+    const filename = `Standard Notes Export - ${application.getArchiveService().formattedDateForExports()}.zip`
+    await downloadBlobOnAndroid(application, zippedDataBlob, filename)
   }
 }
